initial deduplication between user and admin panel
