+++ web/source/panels/admin/index.js
@@ -22,45 +22,13 @@ const Promise = require("bluebird");
 const React = require("react");
 const ReactDom = require("react-dom");
 
-const oauthLib = require("../../lib/oauth.js");
-const Auth = require("./auth");
+const createPanel = require("../lib/panel");
+
 const Settings = require("./settings");
 const Blocks = require("./blocks");
 
 require("../base.css");
 require("./style.css");
-
-function App() {
-	const [oauth, setOauth] = React.useState();
-	const [hasAuth, setAuth] = React.useState(false);
-	const [oauthState, setOauthState] = React.useState(localStorage.getItem("oauth"));
-
-	React.useEffect(() => {
-		let state = localStorage.getItem("oauth");
-		if (state != undefined) {
-			state = JSON.parse(state);
-			let restoredOauth = oauthLib(state.config, state);
-			Promise.try(() => {
-				return restoredOauth.callback();
-			}).then(() => {
-				setAuth(true);
-			});
-			setOauth(restoredOauth);
-		}
-	}, []);
-
-	if (!hasAuth && oauth && oauth.isAuthorized()) {
-		setAuth(true);
-	}
-
-	if (oauth && oauth.isAuthorized()) {
-		return <AdminPanel oauth={oauth} />;
-	} else if (oauthState != undefined) {
-		return "processing oauth...";
-	} else {
-		return <Auth setOauth={setOauth} />;
-	}
-}
 
 function AdminPanel({oauth}) {
 	/* 
@@ -93,4 +61,4 @@ function Logout({oauth}) {
 	);
 }
 
-ReactDom.render(<App/>, document.getElementById("root"));
+createPanel("GoToSocial Admin Panel", ["admin"], AdminPanel);
