+++ web/source/panels/lib/oauth.js
